requirements.txt
anthropic>=0.40.0
click>=8.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
markdownify>=0.13.0
lxml>=5.0.0
weasyprint>=62.0
jinja2>=3.1.0
pydantic>=2.5.0
langdetect>=1.0.9
python-dotenv>=1.0.0
python-slugify>=8.0.0
rich>=13.7.0
